Moved internal package xcube.cli._gen2 to xcube.core.gen2
